+++ src/main/java/org/nrg/containers/services/impl/CommandResolutionServiceImpl.java
@@ -114,6 +114,12 @@ public class CommandResolutionServiceImpl implements CommandResolutionService {
 
     public static final String swarmConstraintsTag = "swarm-constraints";
 
+
+    private static final List<String> supportedParentOutputTypes = Arrays.asList(
+            CommandWrapperOutputEntity.Type.ASSESSOR.getName(),
+            CommandWrapperOutputEntity.Type.SCAN.getName()
+    );
+
     @Autowired
     public CommandResolutionServiceImpl(final CommandService commandService,
                                         final ConfigService configService,
@@ -1932,18 +1938,19 @@ public class CommandResolutionServiceImpl implements CommandResolutionService {
 
                     // Ok, we have found an output. Make sure it can handle another output.
                     // Basically, *this* output handler needs to make a resource, and the
-                    // *target* output handler needs to make an assessor.
+                    // *target* output handler needs to make an assessor or a scan.
                     final boolean thisHandlerIsAResource = commandOutputHandler.type().equals(CommandWrapperOutputEntity.Type.RESOURCE.getName());
-                    final boolean targetHandlerIsAnAssessor = otherOutputHandler.type().equals(CommandWrapperOutputEntity.Type.ASSESSOR.getName());
-                    if (!(thisHandlerIsAResource && targetHandlerIsAnAssessor)) {
+                    final boolean targetHandlerIsSupported = supportedParentOutputTypes.contains(otherOutputHandler.type());
+                    if (!(thisHandlerIsAResource && targetHandlerIsSupported)) {
                         // This output is supposed to be uploaded to an object that is created by another output,
-                        // but that can only happen (as of now, 2018-03-23) when the first output is an assessor
-                        // and any subsequent outputs are resources
+                        // but that can only happen when the first (parent) output is an assessor or a scan
+                        // and any subsequent (child) outputs are resources
                         final String message = String.format("Cannot resolve handler \"%1$s\". " +
                                         "Handler \"%1$s\" has type \"%2$s\"; target handler \"%3$s\" has type \"%4$s\". " +
-                                        "Handler \"%1$s\" must be type Resource, target handler \"%3$s\" needs to be type Assessor.",
+                                        "Handler \"%1$s\" must be type Resource, target handler \"%3$s\" needs to be type %5$s.",
                                 commandOutputHandler.name(), commandOutputHandler.type(),
-                                commandOutputHandler.targetName(), otherOutputHandler.type());
+                                commandOutputHandler.targetName(), otherOutputHandler.type(),
+                                String.join(" OR ", supportedParentOutputTypes));
                         if (Boolean.TRUE.equals(commandOutput.required()) && !outputHasAtLeastOneLegitHandler) {
                             throw new CommandResolutionException(message);
                         } else {